requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
matplotlib
seaborn==0.13.2
flet==0.25.2
ensemble==0.0.dev1
statsmodels==0.14.1
scipy==1.10.1
xgboost==1.7.5
lightgbm==3.3.2
catboost==1.1.1
pyyaml==6.0.2
mlflow
streamlit
imbalanced-learn==0.12.4
-e.